requirements.txt
crewai>=1.9.0
crewai-tools>=1.9.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pypdf>=4.0.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
google-generativeai>=0.5.0
google-genai>=1.0.0
python-multipart>=0.0.9
